+++ heepsilon.core
@@ -14,8 +14,8 @@ filesets:
     - openhwgroup.org:systems:core-v-mini-mcu
     - eslepfl::cgra
     files:
-    - hw/rtl/cgra_x_heep_pkg.sv
-    - hw/rtl/cgra_x_heep_top.sv
+    - hw/rtl/heepsilon_pkg.sv
+    - hw/rtl/heepsilon_top.sv
     file_type: systemVerilogSource
 
   x_heep_system:
@@ -102,7 +102,7 @@ filesets:
     - hw/fpga/sram_wrapper.sv
     - hw/fpga_cgra/cgra_sram_wrapper.sv
     - hw/fpga_cgra/cgra_clock_gate.sv
-    - hw/fpga_cgra/xilinx_cgra_x_heep_wrapper.sv
+    - hw/fpga_cgra/xilinx_heepsilon_wrapper.sv
     file_type: systemVerilogSource
 
   ip-fpga:
@@ -179,7 +179,7 @@ targets:
   default: &default_target
     filesets:
     - files_rtl_generic
-    toplevel: [cgra_x_heep_top]
+    toplevel: [heepsilon_top]
 
   sim:
     <<: *default_target
@@ -277,4 +277,4 @@ targets:
     tools:
       vivado:
         part: xc7z020clg400-1
-    toplevel: [xilinx_cgra_x_heep_wrapper]
+    toplevel: [xilinx_heepsilon_wrapper]
