+++ heepsilon_cfg.hjson
@@ -10,10 +10,10 @@
         // This limit can be increased but manual changes are required
         // Maximum number of columns: 32-log2(rcs_num_instr)-log2(max_columns*rcs_num_instr)
         // Default settings enable: 32-log2(32)-log2(4*32) = 20 columns
-        num_columns: 4
+        num_columns: 11
         // Main impact: more rows equals to more context memory banks (i.e., one per row)
         // There should be no limitation compared to the number of rows compared to columns
-        num_rows: 4
+        num_rows: 5
         // It is possible to limit the maximum number of columns a kernel can use (this saves a bit of ressources)
         // The default value should be the same than num_columns, put an number to change it
         max_columns: default
