+++ sw/applications/cgra_fft/main.c
@@ -7,7 +7,7 @@
 #include "core_v_mini_mcu.h"
 #include "rv_plic.h"
 #include "rv_plic_regs.h"
-#include "cgra_x_heep.h"
+#include "heepsilon.h"
 #include "cgra.h"
 #include "cgra_bitstream.h"
 #include "fxp.h"
